+++ tests/ui/serde/serde.rs
@@ -4,7 +4,7 @@ use derive_where::derive_where;
 
 #[derive_where(Clone)]
 #[serde(crate = "serde_")]
-struct InvalidSerde<T>(PhantomData<T>);
+struct ValidSerde<T>(PhantomData<T>);
 
 #[derive_where(Deserialize)]
 #[serde(bound = "")]