+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+- `#[serde(...)]` attributes are now allowed even when not deriving
+  `De/Serialize` with `#[derive_where(...)]`.
+
 ## [1.5.0] - 2025-06-08
 
 ### Added